+++ docs-src/pages/changelog.svelte
@@ -1,7 +1,7 @@
 <h1 id="changelog">Changelog</h1>
 <h2 id="v631-2023-03-26">v6.3.1 <em>(2023-03-26)</em></h2>
 <ul>
-<li>ButtonGroup styling tweaks (edge buttons padding alignment)</li>
+<li><code>ButtonGroup</code> styling tweaks (edge buttons padding alignment)</li>
 </ul>
 <h2 id="v630-2023-03-26">v6.3.0 <em>(2023-03-26)</em></h2>
 <ul>